ThePhasesOfCraftship/2_best_practice_first/testingBasics/exercises/2_Arrange_Act_Assert_Backwards/2_1_Password_Validator/src/index.ts
export class PasswordValidator {
  constructor() {}

  exec(input: string) {
    const messages = [];
    const hasDigit = !!input.split("").filter((char) => !isNaN(Number(char)))
      .length;
    const hasValidLength = input.length >= 5 && input.length <= 15;
    const hasUpperCaseLetter = input
      .split("")
      .some(
        (char) =>
          typeof char.toLowerCase() !== char && char.toUpperCase() === char
      );

    if (!hasDigit) {
      messages.push(`Your password need include at least one digit: ${input}`);
    }

    if (!hasValidLength) {
      messages.push(
        `Your password length is incorrect (between 5 an 15): ${input}`
      );
    }

    if (!hasUpperCaseLetter) {
      messages.push(
        `Your password need include at least one upper case letter: ${input}`
      );
    }
    return {
      passed: !messages.length,
      messages: messages,
    };
  }
}
